Setup server mux
