+++ cmd/api/main.go
@@ -2,16 +2,12 @@ package main
 
 import (
 	"bookshop/internal/data"
-	"context"
-	"errors"
 	"flag"
-	"github.com/jackc/pgx/v5/pgxpool"
-	_ "github.com/joho/godotenv/autoload"
 	"log/slog"
-	"net/http"
 	"os"
-	"strings"
 	"time"
+
+	_ "github.com/joho/godotenv/autoload"
 )
 
 type config struct {
@@ -33,15 +29,7 @@ type application struct {
 }
 
 func main() {
-	var cfg config
-	flag.StringVar(&cfg.log.level, "log-level", "info", "Logging level (debug|info|warning|error)")
-	flag.StringVar(&cfg.log.format, "log-format", "json", "Logging format (text|json)")
-
-	flag.IntVar(&cfg.db.maxConns, "dbpool-max-conns", 4, "Database max open connections")
-	flag.IntVar(&cfg.db.minConns, "dbpool-min-conns", 1, "Database min idle connections")
-	flag.DurationVar(&cfg.db.maxIdleTime, "dbpool-max-idle-time", 15*time.Minute, "Database max connection idle time")
-
-	flag.Parse()
+	cfg := parseConfig()
 
 	logger := setupLogger(cfg)
 
@@ -58,59 +46,22 @@ func main() {
 		models: data.NewModels(dbpool),
 	}
 
-	err = app.dummyListenAndServe()
+	err = app.run()
 	if err != nil {
 		os.Exit(1)
 	}
 }
 
-func setupDbPool(cfg config) (*pgxpool.Pool, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
+func parseConfig() config {
+	var cfg config
+	flag.StringVar(&cfg.log.level, "log-level", "info", "Logging level (debug|info|warning|error)")
+	flag.StringVar(&cfg.log.format, "log-format", "json", "Logging format (text|json)")
 
-	poolCfg, err := pgxpool.ParseConfig(os.Getenv("DATABASE_URL"))
-	if err != nil {
-		return nil, err
-	}
+	flag.IntVar(&cfg.db.maxConns, "dbpool-max-conns", 4, "Database max open connections")
+	flag.IntVar(&cfg.db.minConns, "dbpool-min-conns", 1, "Database min idle connections")
+	flag.DurationVar(&cfg.db.maxIdleTime, "dbpool-max-idle-time", 15*time.Minute, "Database max connection idle time")
 
-	dbpool, err := pgxpool.NewWithConfig(ctx, poolCfg)
-	if err != nil {
-		return nil, err
-	}
+	flag.Parse()
 
-	err = dbpool.Ping(ctx)
-	if err != nil {
-		return nil, errors.New(err.Error() + ": unable to connect to the database")
-	}
-
-	return dbpool, nil
-}
-
-func setupLogger(cfg config) *slog.Logger {
-	var level slog.Level
-	err := level.UnmarshalText([]byte(cfg.log.level))
-	if err != nil {
-		level = slog.LevelInfo
-	}
-	opts := &slog.HandlerOptions{Level: level}
-
-	switch strings.ToLower(cfg.log.format) {
-	case "text":
-		return slog.New(slog.NewTextHandler(os.Stdout, opts))
-	default:
-		return slog.New(slog.NewJSONHandler(os.Stdout, opts))
-	}
-}
-
-func (app *application) dummyListenAndServe() error {
-	srv := &http.Server{
-		Addr:         ":4000",
-		Handler:      app.routes(),
-		IdleTimeout:  time.Minute,
-		ReadTimeout:  5 * time.Second,
-		WriteTimeout: 10 * time.Second,
-	}
-
-	app.logger.Info("starting server", "addr", srv.Addr)
-	return srv.ListenAndServe()
+	return cfg
 }
